refactored to spotify store; no issue when switching between pages
