+++ src/store/GlobalStore.jsx
@@ -1,8 +1,13 @@
 import * as React from "react";
 import WeatherStore from "./WeatherContext";
+import SpotifyStore from "./SpotifyContext";
 
 const GlobalStore = ({ children }) => {
-  return <WeatherStore>{children}</WeatherStore>;
+  return (
+    <SpotifyStore>
+      <WeatherStore>{children}</WeatherStore>
+    </SpotifyStore>
+  );
 };
 
 export default GlobalStore;
